Do works related to presentation
- Place all presentation related methods and constants into assets.rb
- Require and include assets.rb into game_actions.rb
- Put spaces below game title and below board
- Put a message after user decides to not play again
