--- test.rb
@@ -1,17 +1,0 @@
-a = %q(
-  
- /$$$$$$$$/$$                               /$$$$$$$$                                      /$$$$$$$$                 
-|__  $$__/__/                              |__  $$__/                                     |__  $$__/                 
-   | $$   /$$  /$$$$$$$                       | $$  /$$$$$$   /$$$$$$$                       | $$  /$$$$$$   /$$$$$$ 
-   | $$  | $$ /$$_____/       /$$$$$$         | $$ |____  $$ /$$_____/       /$$$$$$         | $$ /$$__  $$ /$$__  $$
-   | $$  | $$| $$            |______/         | $$  /$$$$$$$| $$            |______/         | $$| $$  \ $$| $$$$$$$$
-   | $$  | $$| $$                             | $$ /$$__  $$| $$                             | $$| $$  | $$| $$_____/
-   | $$  | $$|  $$$$$$$                       | $$|  $$$$$$$|  $$$$$$$                       | $$|  $$$$$$/|  $$$$$$$
-   |__/  |__/ \_______/                       |__/ \_______/ \_______/                       |__/ \______/  \_______/
-                                                                                                                     
-                                                                                                                     
-                                                                                                                     
-
-)
-
-puts a
